Extract ReadOnlyTextCell from widget dlg and use in Help/About
